+++ NEWS.md
@@ -1,2 +1,15 @@
 # Version track
-2020-02-25 version 0.1.0 released
+2020-02-25 
+
+* version 0.1.0 released
+
+2020-05-10
+
+* version 0.2.0 released
+
+*  Update CTLE function to CTLERob function.
+This new function have one more parameter 'rlr_method' which let user choose the robust regression method in 'lmRob','lmrob','ltsReg'.
+
+*  Update class definition of RobMixReg.
+The new class add one slot which return the posterior probability of the mixture regression.
+